# An attempt at "Lazy" JSON parsing