This is an attempt at lazy parsing JSON in hopes of better performance. Below is the entire test output (just the performance tests) but I'll summarize here. The theory is that the lazy parsing shouldn't add any time when decoding the full JSON but can cut down on time when only decoding part of the JSON. I've created another mode `PartialTestModel` that decodes from the same `JSON` as the `TestModel` but only decodes a few properties.

Currently on master, parsing averages to 1.136s and decoding the full model averages 0.733 totaling to **1.869s**. The partial model decodes in 0.117s totaling **1.253s**.

This branch, parsing averages to 0.009s and decoding the full model averages 2.014s totaling to **2.023s**. The partial model decodes in 0.552s totaling **0.561s**.

So you can see that this attempt at lazy JSON adds **8.2%** to the time for full decoding but saves **55%** when decoding the partial model. So worst case is a bit worse but best case could be much better. Seems like this is a positive thing but I want to hear other opinions.

-------------------------------------------------------------------------------------------

This branch:
```
Test Suite 'PerformanceTests' started at 2017-05-12 11:36:35.775
Test Case '-[ArgoTests.PerformanceTests testBigDataDecodesCorrectly]' started.
Test Case '-[ArgoTests.PerformanceTests testBigDataDecodesCorrectly]' passed (2.161 seconds).
Test Case '-[ArgoTests.PerformanceTests testDecodePerformance]' started.
Test Case '-[ArgoTests.PerformanceTests testDecodePerformance]' measured [Time, seconds] average: 2.014, relative standard deviation: 1.569%, values: [2.004167, 2.032448, 2.023439, 2.001006, 1.970307, 2.001316, 2.007409, 2.020993, 2.093751, 1.986349], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[ArgoTests.PerformanceTests testDecodePerformance]' passed (20.509 seconds).
Test Case '-[ArgoTests.PerformanceTests testParsePerformance]' started.
Test Case '-[ArgoTests.PerformanceTests testParsePerformance]' measured [Time, seconds] average: 0.009, relative standard deviation: 10.247%, values: [0.010456, 0.007621, 0.009246, 0.009345, 0.008630, 0.007998, 0.007957, 0.008947, 0.009619, 0.007634], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[ArgoTests.PerformanceTests testParsePerformance]' passed (0.441 seconds).
Test Case '-[ArgoTests.PerformanceTests testPartialDecodePerformance]' started.
Test Case '-[ArgoTests.PerformanceTests testPartialDecodePerformance]' measured [Time, seconds] average: 0.552, relative standard deviation: 3.189%, values: [0.561163, 0.541514, 0.546523, 0.537692, 0.537148, 0.598750, 0.551715, 0.536631, 0.557324, 0.553036], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[ArgoTests.PerformanceTests testPartialDecodePerformance]' passed (5.888 seconds).
Test Suite 'PerformanceTests' passed at 2017-05-12 11:37:04.777.
	 Executed 4 tests, with 0 failures (0 unexpected) in 29.000 (29.002) seconds
```
master:
```
Test Suite 'PerformanceTests' started at 2017-05-12 11:41:59.004
Test Case '-[ArgoTests.PerformanceTests testBigDataDecodesCorrectly]' started.
Test Case '-[ArgoTests.PerformanceTests testBigDataDecodesCorrectly]' passed (2.019 seconds).
Test Case '-[ArgoTests.PerformanceTests testDecodePerformance]' started.
Test Case '-[ArgoTests.PerformanceTests testDecodePerformance]' measured [Time, seconds] average: 0.733, relative standard deviation: 2.405%, values: [0.737361, 0.748237, 0.740130, 0.731497, 0.715472, 0.766685, 0.745634, 0.719830, 0.717083, 0.705184], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[ArgoTests.PerformanceTests testDecodePerformance]' passed (8.834 seconds).
Test Case '-[ArgoTests.PerformanceTests testParsePerformance]' started.
Test Case '-[ArgoTests.PerformanceTests testParsePerformance]' measured [Time, seconds] average: 1.136, relative standard deviation: 1.279%, values: [1.148678, 1.131917, 1.126283, 1.124288, 1.130257, 1.150127, 1.124245, 1.135282, 1.167864, 1.118230], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[ArgoTests.PerformanceTests testParsePerformance]' passed (11.719 seconds).
Test Case '-[ArgoTests.PerformanceTests testPartialDecodePerformance]' started.
Test Case '-[ArgoTests.PerformanceTests testPartialDecodePerformance]' measured [Time, seconds] average: 0.117, relative standard deviation: 3.129%, values: [0.115177, 0.119668, 0.114696, 0.122064, 0.116096, 0.118893, 0.112882, 0.115287, 0.121189, 0.109818], performanceMetricID:com.apple.XCTPerformanceMetric_WallClockTime, baselineName: "", baselineAverage: , maxPercentRegression: 10.000%, maxPercentRelativeStandardDeviation: 10.000%, maxRegression: 0.100, maxStandardDeviation: 0.100
Test Case '-[ArgoTests.PerformanceTests testPartialDecodePerformance]' passed (2.673 seconds).
Test Suite 'PerformanceTests' passed at 2017-05-12 11:42:24.251.
	 Executed 4 tests, with 0 failures (0 unexpected) in 25.245 (25.247) seconds
```