@sharplet 2017-05-12
Are the performance tests against Release configuration?