@tonyd256 2017-05-12
@sharplet yes, Argo tests are setup for release.