@sharplet 2017-05-12
The potential gains here seem worth the trade off to me. It seems like the 8% difference shouldn't make much of a difference on normal sized data, but I don't know if I can back that up.

Is it worth it to try and benchmark with "normal" data and try to quantify the loss? Is there a good way for us to know how frequent it is to leave entire subtrees of a JSON payload unparsed?