@gfontenot 2017-06-01
I'm a little concerned about this leaking out to people writing custom `decode` implementations. I don't really like that they would have to know that they need to perform the `Any -> JSON` conversion manually.

I think this is probably a good start, but I feel like we should be pushing it further. In addition to this, for example, I'd like to see if we can make sure that these decode operations only happen once. This would (I believe) have huge benefits for anyone using our nested object syntax multiple times. For example, `["foo", "bar"]` and `["foo", "baz"]` would decode the full `"foo"` object twice. This seems like the ultimate problem we should be shooting for.